requirements.txt
numpy
Pillow
matplotlib
lpips
# OpenAI CLIP
git+https://github.com/openai/CLIP.git
# Aesthetic PRedictor
aesthetic-predictor-v2-5